requirements.txt
crewai>=0.11.0
python-dotenv>=0.19.0
langchain>=0.1.0
openai>=1.0.0 